實作  一二集各買了一本，第三集買了兩本，價格應為100*3*0.9 + 100 = 370

+++ Hank_TDD_Day2Homework/Models/PoterBook.cs
@@ -9,7 +9,7 @@ namespace Hank_TDD_Day2Homework.Models
     {
         public  PoterVersion Version { get;set; }
 
-        public int Pirce { get { return 100; }  }
+        public double Pirce { get { return 100; }  }
 
     }
